fix: algo and structure

+++ src/base_implementation/algorithms.cpp
@@ -6,7 +6,6 @@
 #include <chrono>
 #include <sycl/sycl.hpp>
 #include "headers/utils.h"
-#include <CL/sycl.hpp>
 #include <iostream>
 
 using namespace sycl;
@@ -27,9 +26,9 @@ void find_communities(HypergraphNotSparse& H) {
     const size_t N = H.num_vertices;
     const size_t E = H.num_hyperedges;
 
-    uint32_t* incidence_matrix_usm = sycl::malloc_shared<uint32_t>(N * E, q);
-    uint32_t* vlabels_usm = sycl::malloc_shared<uint32_t>(N, q);
-    uint32_t* helabels_usm = sycl::malloc_shared<uint32_t>(E, q);
+    uint8_t* incidence_matrix_usm = sycl::malloc_shared<uint8_t>(N * E, q);
+    uint8_t* vlabels_usm = sycl::malloc_shared<uint8_t>(N, q);
+    uint8_t* helabels_usm = sycl::malloc_shared<uint8_t>(E, q);
     size_t* edge_indices_usm = sycl::malloc_shared<size_t>(E, q);
     size_t* vertex_indices_usm = sycl::malloc_shared<size_t>(N, q);
 
@@ -57,18 +56,18 @@ void find_communities(HypergraphNotSparse& H) {
         sycl::event edge_event = q.submit([&](sycl::handler& h) {
             h.parallel_for(range<1>(E), [=](id<1> idx) {
                 size_t e = edge_indices_usm[idx];
-                uint32_t label_counts[128] = {0};
+                uint8_t label_counts[128] = {0};
 
                 for (size_t v = 0; v < N; ++v) {
                     if (incidence_matrix_usm[v * E + e] == 1) {
-                        uint32_t lbl = vlabels_usm[v];
-                        if (lbl < 128 && lbl != std::numeric_limits<uint32_t>::max()) {
+                        uint8_t lbl = vlabels_usm[v];
+                        if (lbl < 128 && lbl != std::numeric_limits<uint8_t>::max()) {
                             label_counts[lbl]++;
                         }
                     }
                 }
 
-                uint32_t max_count = 0, best_label = std::numeric_limits<uint32_t>::max();
+                uint8_t max_count = 0, best_label = std::numeric_limits<uint8_t>::max();
                 for (size_t i = 0; i < 128; ++i) {
                     if (label_counts[i] > max_count) {
                         max_count = label_counts[i];
@@ -76,7 +75,7 @@ void find_communities(HypergraphNotSparse& H) {
                     }
                 }
 
-                if (best_label != std::numeric_limits<uint32_t>::max())
+                if (best_label != std::numeric_limits<uint8_t>::max())
                     helabels_usm[e] = best_label;
             });
         });
@@ -92,18 +91,18 @@ void find_communities(HypergraphNotSparse& H) {
         sycl::event vertex_event = q.submit([&](sycl::handler& h) {
             h.parallel_for(range<1>(N), [=](id<1> idx) {
                 size_t v = vertex_indices_usm[idx];
-                uint32_t label_counts[128] = {0};
+                uint8_t label_counts[128] = {0};
 
                 for (size_t e = 0; e < E; ++e) {
                     if (incidence_matrix_usm[v * E + e] == 1) {
-                        uint32_t lbl = helabels_usm[e];
-                        if (lbl < 128 && lbl != std::numeric_limits<uint32_t>::max()) {
+                        uint8_t lbl = helabels_usm[e];
+                        if (lbl < 128 && lbl != std::numeric_limits<uint8_t>::max()) {
                             label_counts[lbl]++;
                         }
                     }
                 }
 
-                uint32_t max_count = 0, best_label = vlabels_usm[v];
+                uint8_t max_count = 0, best_label = vlabels_usm[v];
                 for (size_t i = 0; i < 128; ++i) {
                     if (label_counts[i] > max_count) {
                         max_count = label_counts[i];
@@ -111,7 +110,7 @@ void find_communities(HypergraphNotSparse& H) {
                     }
                 }
 
-                if (vlabels_usm[v] != best_label && best_label != std::numeric_limits<uint32_t>::max()) {
+                if (vlabels_usm[v] != best_label && best_label != std::numeric_limits<uint8_t>::max()) {
                     vlabels_usm[v] = best_label;
                     *stop_flag_device = false;
                 }
@@ -139,9 +138,9 @@ void find_communities(HypergraphNotSparse& H) {
     sycl::free(vertex_indices_usm, q);
 }
 
-// void transpose_incidence_matrix(sycl::queue& q, const std::vector<std::vector<uint32_t>>& incidence_matrix, uint32_t* incidence_matrix_T, size_t N, size_t E) {
+// void transpose_incidence_matrix(sycl::queue& q, const std::vector<std::vector<uint8_t>>& incidence_matrix, uint8_t* incidence_matrix_T, size_t N, size_t E) {
 
-//     uint32_t* incidence_flat = sycl::malloc_shared<uint32_t>(N * E, q);
+//     uint8_t* incidence_flat = sycl::malloc_shared<uint8_t>(N * E, q);
 
 //     for (size_t v = 0; v < N; ++v) {
 //         for (size_t e = 0; e < E; ++e) {
@@ -164,8 +163,8 @@ void find_communities(HypergraphNotSparse& H) {
 //     sycl::free(incidence_flat, q);
 // }
 
-void transpose_incidence_matrix(sycl::queue& q, const std::vector<std::vector<uint32_t>>& incidence_matrix, uint32_t* incidence_matrix_T, size_t N, size_t E) {
-    uint32_t* incidence_flat = sycl::malloc_shared<uint32_t>(N * E, q);
+void transpose_incidence_matrix(sycl::queue& q, const std::vector<std::vector<uint8_t>>& incidence_matrix, uint8_t* incidence_matrix_T, size_t N, size_t E) {
+    uint8_t* incidence_flat = sycl::malloc_shared<uint8_t>(N * E, q);
 
     for (size_t v = 0; v < N; ++v) {
         for (size_t e = 0; e < E; ++e) {
@@ -176,7 +175,7 @@ void transpose_incidence_matrix(sycl::queue& q, const std::vector<std::vector<ui
     auto start_time = std::chrono::high_resolution_clock::now();
 
     q.submit([&](sycl::handler& h) {
-        sycl::local_accessor<uint32_t, 2> tile(sycl::range<2>(TILE_SIZE, TILE_SIZE), h);
+        sycl::local_accessor<uint8_t, 2> tile(sycl::range<2>(TILE_SIZE, TILE_SIZE), h);
 
         h.parallel_for(sycl::nd_range<2>(
                            sycl::range<2>((E + TILE_SIZE - 1) / TILE_SIZE * TILE_SIZE,
@@ -213,8 +212,15 @@ void find_communities_transpose(HypergraphNotSparse& H) {
     sycl::queue q;
     try {
         sycl::queue q(sycl::gpu_selector_v);
-        std::cout << "Selected device: "
-                << q.get_device().get_info<sycl::info::device::name>() << std::endl;
+        std::cout << "Selected device: " << q.get_device().get_info<sycl::info::device::name>() << std::endl;
+        // std::cout << "Backend: "
+        //   << q.get_device().get_platform().get_info<sycl::info::platform::name>() << std::endl;
+        // auto dev = q.get_device();
+        // auto platform = dev.get_platform();
+        // std::cout << "Device: " << dev.get_info<sycl::info::device::name>() << "\n";
+        // std::cout << "Vendor: " << dev.get_info<sycl::info::device::vendor>() << "\n";
+        // std::cout << "Platform: " << platform.get_info<sycl::info::platform::name>() << "\n";
+        // std::cout << "Backend: " << platform.get_info<sycl::info::platform::vendor>() << "\n";
     } catch (sycl::exception const& e) {
         std::cerr << "Failed to create SYCL queue: " << e.what() << std::endl;
         std::exit(EXIT_FAILURE);
@@ -223,9 +229,9 @@ void find_communities_transpose(HypergraphNotSparse& H) {
     const size_t N = H.num_vertices;
     const size_t E = H.num_hyperedges;
 
-    uint32_t* incidence_matrix_T = sycl::malloc_shared<uint32_t>(E * N, q);
-    uint32_t* vlabels_usm = sycl::malloc_shared<uint32_t>(N, q);
-    uint32_t* helabels_usm = sycl::malloc_shared<uint32_t>(E, q);
+    uint8_t* incidence_matrix_T = sycl::malloc_shared<uint8_t>(E * N, q);
+    uint8_t* vlabels_usm = sycl::malloc_shared<uint8_t>(N, q);
+    uint8_t* helabels_usm = sycl::malloc_shared<uint8_t>(E, q);
     size_t* edge_indices_usm = sycl::malloc_shared<size_t>(E, q);
     size_t* vertex_indices_usm = sycl::malloc_shared<size_t>(N, q);
 
@@ -249,18 +255,18 @@ void find_communities_transpose(HypergraphNotSparse& H) {
         sycl::event edge_event = q.submit([&](sycl::handler& h) {
             h.parallel_for(sycl::range<1>(E), [=](sycl::id<1> idx) {
                 size_t e = edge_indices_usm[idx];
-                uint32_t label_counts[128] = {0};
+                uint8_t label_counts[128] = {0};
 
                 for (size_t v = 0; v < N; ++v) {
                     if (incidence_matrix_T[e * N + v] == 1) {
-                        uint32_t lbl = vlabels_usm[v];
-                        if (lbl < 128 && lbl != std::numeric_limits<uint32_t>::max()) {
+                        uint8_t lbl = vlabels_usm[v];
+                        if (lbl < 128 && lbl != std::numeric_limits<uint8_t>::max()) {
                             label_counts[lbl]++;
                         }
                     }
                 }
 
-                uint32_t max_count = 0, best_label = std::numeric_limits<uint32_t>::max();
+                uint8_t max_count = 0, best_label = std::numeric_limits<uint8_t>::max();
                 for (size_t i = 0; i < 128; ++i) {
                     if (label_counts[i] > max_count) {
                         max_count = label_counts[i];
@@ -268,7 +274,7 @@ void find_communities_transpose(HypergraphNotSparse& H) {
                     }
                 }
 
-                if (best_label != std::numeric_limits<uint32_t>::max())
+                if (best_label != std::numeric_limits<uint8_t>::max())
                     helabels_usm[e] = best_label;
             });
         });
@@ -284,18 +290,18 @@ void find_communities_transpose(HypergraphNotSparse& H) {
         sycl::event vertex_event = q.submit([&](sycl::handler& h) {
             h.parallel_for(sycl::range<1>(N), [=](sycl::id<1> idx) {
                 size_t v = vertex_indices_usm[idx];
-                uint32_t label_counts[128] = {0};
+                uint8_t label_counts[128] = {0};
 
                 for (size_t e = 0; e < E; ++e) {
                     if (incidence_matrix_T[e * N + v] == 1) {
-                        uint32_t lbl = helabels_usm[e];
-                        if (lbl < 128 && lbl != std::numeric_limits<uint32_t>::max()) {
+                        uint8_t lbl = helabels_usm[e];
+                        if (lbl < 128 && lbl != std::numeric_limits<uint8_t>::max()) {
                             label_counts[lbl]++;
                         }
                     }
                 }
 
-                uint32_t max_count = 0, best_label = vlabels_usm[v];
+                uint8_t max_count = 0, best_label = vlabels_usm[v];
                 for (size_t i = 0; i < 128; ++i) {
                     if (label_counts[i] > max_count) {
                         max_count = label_counts[i];
@@ -303,7 +309,7 @@ void find_communities_transpose(HypergraphNotSparse& H) {
                     }
                 }
 
-                if (vlabels_usm[v] != best_label && best_label != std::numeric_limits<uint32_t>::max()) {
+                if (vlabels_usm[v] != best_label && best_label != std::numeric_limits<uint8_t>::max()) {
                     vlabels_usm[v] = best_label;
                     *stop_flag_device = false;
                 }
