+++ src/base_implementation/headers/utils.h
@@ -9,10 +9,10 @@ struct HypergraphNotSparse
     std::size_t num_vertices;
     std::size_t num_hyperedges;
     
-    std::vector<std::vector<std::uint32_t>> incidence_matrix;
+    std::vector<std::vector<std::uint8_t>> incidence_matrix;
 
-    std::vector<std::uint32_t> vertex_labels;
-    std::vector<std::uint32_t> hyperedge_labels;
+    std::vector<std::uint8_t> vertex_labels;
+    std::vector<std::uint8_t> hyperedge_labels;
 };
 
 HypergraphNotSparse generate_hypergraph(std::size_t N, std::size_t E, double p);
